start of HW2
-Added a few more servlets we will need - most are empty currently.
-Added User class to session.  A new user class was created as well.
The new user class is basically a container class to better accommodate
java beans.  It is in the flight.bizlogic package.
-We may want to rename the current "Users" class to something like
LoginHelper or UserHelper (and also place into a package) to make it
less confusing with the current class.  We will have some more work to
do with this class for corner cases, such as an incorrect password
attempt.
-divErrorMessage section added to Logic.jsp.  This will be where we
display any errors in logging in, such as an incorrect user/login.  Will
also be used to display an error message if the user tries to access a
page when not logged in.
-look at FlightSearchResults servlet and ViewAndBook.jsp pages for java
bean example implementation
-FlightRecord class updated for bean use.  Depreciated functions are
marked, please be sure not to use these as I'll be removing these later
on.

-we may need to rethink the jquery for including pages (sidebar) and
switch it to a JSP directive, or change some settings for how it is
being loaded.  It is currently caching the whole sidebar and will not
load the new version unless manually refreshed.  This is a problem when
we add the user's display to the sidebar as it will not update after the
user logs in.